+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.1.0] - 2023-02-07
+
+### Added
+
+- Following functionality added:
+  - Open project
+  - Save project
+  - Close project
+  - New project
+
+- Following data retrievable
+  - Has scene been modified
+  - Current scene name
+  - Current scene file path
+
 ## [0.0.1] - 2023-02-07
 
 ### Added
@@ -14,3 +29,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [0.0.1]: https://github.com/LisaGG89/pydobe/releases/tag/v0.0.1
+[0.1.0]: https://github.com/LisaGG89/pydobe/releases/tag/v0.1.0